pyproject.toml
[tool.poetry]
name = "48ix-stats"
version = "0.0.1"
description = "48-IX Statistics API"
authors = ["checktheroads <matt@48ix.net>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.7"
fastapi = "^0"
httpx = "^0"
pydantic = "^1.5"
loguru = "^0.4"
pyyaml = "^5.3"
pendulum = "^2.1"
click = "^7.1"
rich = "^3.0"
uvicorn = "^0.11.5"

[tool.poetry.dev-dependencies]
black = "^19.10b0"
flake8 = "^3.7.9"
isort = "^4.3.21"
flake8-black = "^0.1.1"
flake8-isort = "^3.0.0"
bandit = "^1.6.2"
flake8-bandit = "^2.1.2"
flake8-deprecated = "^1.3"
flake8-docstrings = "^1.5.0"
pep8-naming = "^0.10.0"
stackprinter = "^0.2.3"
pre-commit = "^2.6.0"

[tool.black]
line-length = 88

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"

